requirements.txt
# Core ML/DL frameworks
# torch==2.5.1+cu118
# torchvision==0.20.1+cu118
# torchaudio==2.5.1+cu118
# xformers==0.0.28.post3
pytorch_lightning==2.4.0
timm==1.0.20

# Computer Vision and 3D Processing
opencv-python>=4.8.1.78
open3d==0.17.0
kornia==0.7.4
scikit-image==0.24.0
trimesh==4.5.2
plyfile==1.1
imageio==2.36.0

# Scientific Computing
numpy==1.26.3
scipy==1.14.1
matplotlib==3.9.2
e3nn==0.5.1

# Image/Video Processing
moviepy==1.0.3
sk-video==1.1.10
lpips==0.1.4
colorspacious==1.1.2

# Configuration and Utilities
hydra-core==1.3.2
omegaconf==2.3.0
einops==0.8.0
tqdm==4.65.2
pyyaml==6.0.2
dacite==1.8.1
jaxtyping==0.2.34
beartype==0.19.0

# Development tools
black==24.10.0
ruff==0.11.4
wheel==0.44.0
colorama==0.4.6

# Visualization and UI
tabulate==0.9.0
svg.py==1.5.0

# Monitoring and logging
wandb==0.18.5

# Custom packages
git+https://github.com/dcharatan/diff-gaussian-rasterization-modified
